Add files via upload
php_web is a zip file which contains .php files, .sql (database file) and other css file, JavaScript file and other bootstrap buitin files.
For the execution, you need to run the localhost and the file name. in my case, my file name is php_web. So, you need to write the URL (localhost//php_web/) on the browser.
But the very important thing is to copy this file to the htdocs folder of your downloaded server.
XAMPP is the server I used, 
https://xampp-windows.en.softonic.com/download